# openapi.json
{"openapi": "3.1.0", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/project/root": {"get": {"tags": ["project"], "summary": "Get Root", "operationId": "get_root_folder", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "string", "title": "Response Get Root Folder"}}}}}}, "put": {"tags": ["project"], "summary": "Update Root", "operationId": "update_root_folder", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProjectUpdateRoot"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/project": {"post": {"tags": ["project"], "summary": "Create Project", "operationId": "create_project", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProjectCreationArgs"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "string", "title": "Response Create Project"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/project/{project_id}": {"delete": {"tags": ["project"], "summary": "Delete Project", "operationId": "delete_project", "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Project Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/project/{project_id}/meta": {"put": {"tags": ["project"], "summary": "Update Project Metadata", "operationId": "update_project_metadata", "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Project Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProjectUpdateMetadata"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "get": {"tags": ["project"], "summary": "Get Project Metadata", "operationId": "get_project_metadata", "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Project Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProjectMetadata"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/project/all": {"get": {"tags": ["project"], "summary": "Get All Projects", "operationId": "get_all_projects", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/ProjectMetadata"}, "type": "array", "title": "Response Get All Projects"}}}}}}}, "/runs/create": {"post": {"tags": ["runs"], "summary": "Create Run", "operationId": "create_run", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/RunCreationArgs"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/runs/all": {"get": {"tags": ["runs"], "summary": "Get All Runs", "operationId": "get_all_runs", "parameters": [{"name": "project_id", "in": "query", "required": true, "schema": {"type": "string", "title": "Project Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/Run"}, "title": "Response Get All Runs"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/runs/{project_id}/{run_name}/data": {"post": {"tags": ["runs"], "summary": "Get Run Data", "operationId": "get_run_data", "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Project Id"}}, {"name": "run_name", "in": "path", "required": true, "schema": {"type": "string", "title": "Run Name"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/runs/{project_id}/{run_name}/config": {"put": {"tags": ["runs"], "summary": "Update Run Config", "operationId": "update_run_config", "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Project Id"}}, {"name": "run_name", "in": "path", "required": true, "schema": {"type": "string", "title": "Run Name"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/runs": {"delete": {"tags": ["runs"], "summary": "Delete Run", "operationId": "delete_run", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/RunDeletionArgs"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/runs/{project_id}/{run_name}/default": {"get": {"tags": ["runs"], "summary": "Get Default Config", "operationId": "get_default_config", "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Project Id"}}, {"name": "run_name", "in": "path", "required": true, "schema": {"type": "string", "title": "Run Name"}}, {"name": "config_type", "in": "query", "required": true, "schema": {"type": "string", "title": "Config Type"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PPOAgentControllerConfigModel"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/sessions/start": {"post": {"tags": ["session"], "summary": "Start New Session", "operationId": "start_new_session", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/SessionStartArgs"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Session"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/sessions/{session_id}/stop": {"post": {"tags": ["session"], "summary": "Stop Session", "operationId": "stop_session", "parameters": [{"name": "session_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Session Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/sessions/all": {"post": {"tags": ["session"], "summary": "Get All Sessions", "operationId": "get_all_sessions", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/SessionGetAllArgs"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/Session"}, "type": "array", "title": "Response Get All Sessions"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/sessions/{session_id}/health": {"post": {"tags": ["session"], "summary": "Get Session Health", "operationId": "get_session_health", "parameters": [{"name": "session_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Session Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SessionGetHealth"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "string", "title": "Response Get Session Health"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/": {"get": {"summary": "Ping", "operationId": "ping__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"BaseModel": {"properties": {}, "type": "object", "title": "BaseModel"}, "ExperienceBufferConfigModel": {"properties": {"max_size": {"type": "integer", "title": "Max Size", "default": 100000}, "device": {"type": "string", "enum": ["cpu", "cuda", "ipu", "xpu", "mkldnn", "opengl", "opencl", "ideep", "hip", "ve", "fpga", "maia", "xla", "lazy", "vulkan", "mps", "meta", "hpu", "mtia", "privateuseone"], "title": "Device", "default": "auto"}, "save_experience_buffer_in_checkpoint": {"type": "boolean", "title": "Save Experience Buffer In Checkpoint", "default": true}, "trajectory_processor_config": {"additionalProperties": true, "type": "object", "title": "Trajectory Processor Config"}}, "additionalProperties": false, "type": "object", "title": "ExperienceBufferConfigModel"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "LogConfig": {"properties": {"stdout": {"type": "string", "title": "Stdout"}, "stderr": {"type": "string", "title": "Stderr"}}, "type": "object", "required": ["stdout", "stderr"], "title": "LogConfig"}, "PPOAgentControllerConfigModel": {"properties": {"timesteps_per_iteration": {"type": "integer", "title": "Timesteps Per Iteration", "default": 50000}, "save_every_ts": {"type": "integer", "title": "Save Every Ts", "default": 1000000}, "run_suffix": {"type": "string", "title": "Run Suffix"}, "checkpoint_load_folder": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Checkpoint Load Folder"}, "n_checkpoints_to_keep": {"type": "integer", "title": "N Checkpoints To Keep", "default": 5}, "random_seed": {"type": "integer", "title": "Random Seed", "default": 123}, "save_mid_iteration_data_in_checkpoint": {"type": "boolean", "title": "Save Mid Iteration Data In Checkpoint", "default": true}, "learner_config": {"$ref": "#/components/schemas/PPOLearnerConfigModel"}, "experience_buffer_config": {"$ref": "#/components/schemas/ExperienceBufferConfigModel"}, "run_name": {"type": "string", "title": "Run Name", "default": "rlgym-learn-run"}, "metrics_logger_config": {"anyOf": [{"$ref": "#/components/schemas/BaseModel"}, {"type": "null"}], "title": "Metrics Logger Config"}}, "additionalProperties": false, "type": "object", "title": "PPOAgentControllerConfigModel"}, "PPOLearnerConfigModel": {"properties": {"dtype": {"type": "string", "enum": ["bits16", "uint1", "quint2x4", "quint4x2", "int64", "uint16", "int4", "uint3", "int16", "uint32", "int8", "uint2", "float4_e2m1fn_x2", "uint8", "int3", "int5", "float64", "complex32", "float16", "uint4", "int1", "float8_e4m3fn", "float8_e8m0fnu", "uint64", "int32", "float8_e4m3fnuz", "bits1x8", "bits2x4", "qint32", "complex128", "float32", "complex64", "bits8", "int6", "int2", "float8_e5m2", "uint5", "quint8", "bits4x2", "bfloat16", "float8_e5m2fnuz", "qint8", "uint7", "uint6", "bool", "int7"], "title": "Dtype", "default": "float32"}, "n_epochs": {"type": "integer", "title": "N Epochs", "default": 1}, "batch_size": {"type": "integer", "title": "Batch Size", "default": 50000}, "n_minibatches": {"type": "integer", "title": "N Minibatches", "default": 1}, "ent_coef": {"type": "number", "title": "Ent Coef", "default": 0.005}, "clip_range": {"type": "number", "title": "Clip Range", "default": 0.2}, "actor_lr": {"type": "number", "title": "Actor Lr", "default": 0.0003}, "critic_lr": {"type": "number", "title": "Critic Lr", "default": 0.0003}, "advantage_normalization": {"type": "boolean", "title": "Advantage Normalization", "default": true}, "device": {"type": "string", "enum": ["cpu", "cuda", "ipu", "xpu", "mkldnn", "opengl", "opencl", "ideep", "hip", "ve", "fpga", "maia", "xla", "lazy", "vulkan", "mps", "meta", "hpu", "mtia", "privateuseone"], "title": "Device", "default": "auto"}, "cudnn_benchmark_mode": {"type": "boolean", "title": "Cudnn Benchmark Mode", "default": true}}, "additionalProperties": false, "type": "object", "title": "PPOLearnerConfigModel"}, "ProjectCreationArgs": {"properties": {"name": {"type": "string", "title": "Name"}, "interpreter": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Interpreter"}}, "type": "object", "required": ["name"], "title": "ProjectCreationArgs"}, "ProjectMetadata": {"properties": {"name": {"type": "string", "title": "Name"}, "id": {"type": "string", "title": "Id"}, "interpreter": {"type": "string", "title": "Interpreter"}}, "type": "object", "required": ["name", "id", "interpreter"], "title": "ProjectMetadata"}, "ProjectUpdateMetadata": {"properties": {"name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Name"}, "interpreter": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Interpreter"}}, "type": "object", "title": "ProjectUpdateMetadata"}, "ProjectUpdateRoot": {"properties": {"path": {"type": "string", "title": "Path"}}, "type": "object", "required": ["path"], "title": "ProjectUpdateRoot"}, "Run": {"properties": {"project_id": {"type": "string", "title": "Project Id"}, "name": {"type": "string", "title": "Name"}}, "type": "object", "required": ["project_id", "name"], "title": "Run"}, "RunCreationArgs": {"properties": {"run_name": {"type": "string", "title": "Run Name"}, "project_id": {"type": "string", "title": "Project Id"}}, "type": "object", "required": ["run_name", "project_id"], "title": "RunCreationArgs"}, "RunDeletionArgs": {"properties": {"run_name": {"type": "string", "title": "Run Name"}, "project_id": {"type": "string", "title": "Project Id"}}, "type": "object", "required": ["run_name", "project_id"], "title": "RunDeletionArgs"}, "Session": {"properties": {"session_id": {"type": "string", "title": "Session Id"}, "project_id": {"type": "string", "title": "Project Id"}, "run_name": {"type": "string", "title": "Run Name"}, "status": {"type": "string", "title": "Status"}, "logs": {"$ref": "#/components/schemas/LogConfig"}}, "type": "object", "required": ["session_id", "project_id", "run_name", "status", "logs"], "title": "Session"}, "SessionGetAllArgs": {"properties": {"run_name": {"type": "string", "title": "Run Name"}, "project_id": {"type": "string", "title": "Project Id"}}, "type": "object", "required": ["run_name", "project_id"], "title": "SessionGetAllArgs"}, "SessionGetHealth": {"properties": {"run_name": {"type": "string", "title": "Run Name"}, "project_id": {"type": "string", "title": "Project Id"}}, "type": "object", "required": ["run_name", "project_id"], "title": "SessionGetHealth"}, "SessionStartArgs": {"properties": {"run_name": {"type": "string", "title": "Run Name"}, "project_id": {"type": "string", "title": "Project Id"}}, "type": "object", "required": ["run_name", "project_id"], "title": "SessionStartArgs"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}